Don't run volFixAll on multi-component meshes
Multi-component meshes generally require manual intervention by the researcher.
They should not be fixed automatically because a multi-component mesh likely
came from contours that were misaligned or traced incorrectly in Reconstruct.

Multi-component meshes are typically the result of badly aligned contours in
Reconstruct, or contours that don't overlap enough from one section to the
next.  They can also result from reversal of handedness of contours.

+++ processor_tool.py
@@ -892,7 +892,7 @@ class ProcessorToolSceneProperty(bpy.types.PropertyGroup):
             if mesh_props.components >1:           
                 print('\nFound Multi-component Mesh: %s\n' % (obj.name))
                 self.include_list[name].multi_component = True
-            if (mesh_props.manifold == False) or (mesh_props.watertight == False) or (mesh_props.normal_status == 'Inconsistent Normals'): 
+            if ((mesh_props.manifold == False) or (mesh_props.watertight == False) or (mesh_props.normal_status == 'Inconsistent Normals') and mesh_props.components == 1): 
                 print('\nFixing Single Flawed Mesh: %s\n' % (contour_name))
                 name = obj.name
                 m = obj.data
@@ -946,7 +946,7 @@ class ProcessorToolSceneProperty(bpy.types.PropertyGroup):
                 if mesh_props.components >1:
                     print('\nFound Multi-component Mesh: %s\n' % (obj.name))
                     self.include_list[name].multi_component = True
-                if (mesh_props.manifold == False) or (mesh_props.watertight == False) or (mesh_props.normal_status == 'Inconsistent Normals'): 
+                if ((mesh_props.manifold == False) or (mesh_props.watertight == False) or (mesh_props.normal_status == 'Inconsistent Normals') and mesh_props.components == 1): 
                     print('\nFound Flawed Mesh: %s\n' % (obj.name))
                     m = obj.data
                     context.scene.objects.unlink(obj)
